Moving util class to internal package
